Refactor Project List Page
